audiobook-rss/requirements.txt
feedgen==0.9.0
mutagen==1.45.1
python-dateutil==2.8.1
PyYAML==5.4.1
